serialization: move under src/, prune cereal-includes
- relocate include/vierkant/serialization -> src/serialization; include/ is
  cereal-free by construction, so the install-rule needs no exclusion
- split the JSON specializations into optional_nvp_json.hpp, keeping rapidjson
  out of the binary bundle-path (bundle.cpp: 986 -> 947 headers, 37 -> 0)
- add types/polymorphic.hpp for CEREAL_REGISTER_TYPE, which only ever worked
  via a side-effect include of types/memory.hpp
- give components.hpp and collision_cereal.hpp the type-supports they were
  borrowing from siblings; every header now compiles standalone
- drop the unused type-support includes
